moving grid to src/
